+++ .eslintrc
@@ -11,6 +11,7 @@
     "rules": {
         "simple-import-sort/imports": "error",
         "import/extensions": 0,
+        "import/export": 0,
         "comma-dangle": 0,
         "no-plusplus": 0,
         "no-continue": 0,
@@ -27,6 +28,7 @@
         "react/jsx-filename-extension": "off",
         "react/prefer-stateless-function": "off",
         "no-underscore-dangle": "off",
+        "import/no-cycle": "off",
         "import/no-extraneous-dependencies": [
             "error",
             {
