+++ src/constants/index.js
@@ -1,1 +1,7 @@
+export * from './buttons';
 export * from './componentsConstants';
+export * from './fakeLanguagesData';
+export * from './fakeThemeData';
+export * from './links';
+export * from './pages';
+export * from './sizes';
